clipboard works on os x
